\name{WansbeekMeijer}
\docType{data}
\alias{WansbeekMeijer}
\alias{NetherlandsTV}

\title{Factor Example from Wansbeek and Meijer}

\description{
	Netherlands TV viewership example p 171, Wansbeek and Meijer (2000) 
}
\usage{
	data(WansbeekMeijer)
}
\details{
	The object NetherlandsTV is loaded from the data file WansbeekMeijer.
}
\format{
	The object NetherlandsTV is a correlation matrix.
}
\source{
	Tom Wansbeek and Erik Meijer (2000) \emph{Measurement 
   Error and Latent Variables in Econometrics}, Amsterdam: North-Holland.
}
\seealso{
	\code{\link{GPForth}},
	\code{\link{Thurstone}},
	\code{\link{Harman}}
}
\keyword{datasets}
