\name{NormalizingWeight}
\alias{NormalizingWeight}

\title{Internal Utility for Normalizing Weights}
\usage{
    NormalizingWeight(A, normalize=FALSE)
    }
\arguments{
    \item{A}{A loading matrix.}
    \item{normalize}{An indication of if/how the matrix should be normalized.}
}

\description{
   See GPFobliq and GPForth.
   }
\value{A matrix.}

\keyword{internal}

